src/python/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.0
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
idna==3.4
Jinja2==3.1.2
lit==16.0.6
llvmlite==0.40.1
Marketingtool @ file:///home/robertzeng/project/toyproject/social_market/src/python
MarkupSafe==2.1.3
more-itertools==10.0.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
regex==2023.6.3
requests==2.31.0
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
